Iterating on some feedback.
